backend/requirements.txt
flask
flask-cors
opencv-python
numpy
matplotlib
Pillow
